Some code has been commented out so that it can be loaded normally in 5.3.2, and these codes will be corrected later, In addition, in the included directory, the python directory is currently in hard code and needs to be changed back later

+++ Source/UnrealEnginePython/Private/Wrappers/UEPyFSoftSkinVertex.cpp
@@ -160,7 +160,7 @@ static int py_ue_fsoft_skin_vertex_set_influence_bones(ue_PyFSoftSkinVertex *sel
 
 static PyObject *py_ue_fsoft_skin_vertex_get_influence_weights(ue_PyFSoftSkinVertex *self, void *closure)
 {
-	uint8 *data = self->ss_vertex.InfluenceWeights;
+	uint16*data = self->ss_vertex.InfluenceWeights;
 	return Py_BuildValue((char*)"(iiiiiiii)", data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7]);
 }
